Moves files/items recursive out of an directory structure to a new location.
Moves files/items recursive out of an directory structure to a new
location.
The existing folder structure will be kept in the destination.

Assume you want to move all playlists from a music dir to a new
location. The folder structure in the new location should be the same as
in source, but there only playlist-files should be placed there.